feat(cli): Add `--init` command for interactive configuration setup
