+++ .repopackignore
@@ -1,4 +1,4 @@
 node_modules
 .yarn
 .eslinttcache
-tests/fixtures
+tests/integration-tests/fixtures
