+++ src/cli/cliRunner.ts
@@ -1,22 +1,13 @@
-import path from 'node:path';
 import process from 'node:process';
 import { program, OptionValues } from 'commander';
-import pc from 'picocolors';
-import { PackResult, pack } from '../core/packager.js';
-import {
-  RepopackConfigCli,
-  RepopackConfigFile,
-  RepopackConfigMerged,
-  RepopackOutputStyle,
-} from '../config/configTypes.js';
-import { loadFileConfig, mergeConfigs } from '../config/configLoader.js';
-import { logger } from '../shared/logger.js';
+import { RepopackOutputStyle } from '../config/configTypes.js';
 import { getVersion } from '../core/file/packageJsonParser.js';
 import { handleError } from '../shared/errorHandler.js';
-import Spinner from './cliSpinner.js';
-import { printSummary, printTopFiles, printCompletion, printSecurityCheck } from './cliPrinter.js';
+import { runInitCommand } from './commands/initCommandRunner.js';
+import { runVersionCommand } from './commands/versionCommandRunner.js';
+import { runDefaultCommand } from './commands/defaultCommandRunner.js';
 
-interface CliOptions extends OptionValues {
+export interface CliOptions extends OptionValues {
   version?: boolean;
   output?: string;
   include?: string;
@@ -26,6 +17,7 @@ interface CliOptions extends OptionValues {
   topFilesLen?: number;
   outputShowLineNumbers?: boolean;
   style?: RepopackOutputStyle;
+  init?: boolean;
 }
 
 export async function run() {
@@ -45,6 +37,7 @@ export async function run() {
       .option('--output-show-line-numbers', 'add line numbers to each line in the output')
       .option('--style <type>', 'specify the output style (plain or xml)')
       .option('--verbose', 'enable verbose logging for detailed output')
+      .option('--init', 'initialize a new repopack.config.json file')
       .action((directory = '.', options: CliOptions) => executeAction(directory, process.cwd(), options));
 
     await program.parseAsync(process.argv);
@@ -53,87 +46,16 @@ export async function run() {
   }
 }
 
-const executeAction = async (directory: string, rootDir: string, options: CliOptions) => {
-  const version = await getVersion();
-
+const executeAction = async (directory: string, cwd: string, options: CliOptions) => {
   if (options.version) {
-    console.log(version);
+    await runVersionCommand();
     return;
   }
 
-  console.log(pc.dim(`\n📦 Repopack v${version}\n`));
-
-  logger.setVerbose(options.verbose || false);
-  logger.trace('Loaded CLI options:', options);
-
-  // Load the config file
-  const fileConfig: RepopackConfigFile = await loadFileConfig(rootDir, options.config ?? null);
-  logger.trace('Loaded file config:', fileConfig);
-
-  // Parse the CLI options
-  const cliConfig: RepopackConfigCli = {};
-  if (options.output) {
-    cliConfig.output = { filePath: options.output };
-  }
-  if (options.include) {
-    cliConfig.include = options.include.split(',');
-  }
-  if (options.ignore) {
-    cliConfig.ignore = { customPatterns: options.ignore.split(',') };
-  }
-  if (options.topFilesLen !== undefined) {
-    cliConfig.output = { ...cliConfig.output, topFilesLength: options.topFilesLen };
-  }
-  if (options.outputShowLineNumbers !== undefined) {
-    cliConfig.output = { ...cliConfig.output, showLineNumbers: options.outputShowLineNumbers };
-  }
-  if (options.style) {
-    cliConfig.output = { ...cliConfig.output, style: options.style.toLowerCase() as RepopackOutputStyle };
-  }
-  logger.trace('CLI config:', cliConfig);
-
-  // Merge default, file, and CLI configs
-  const config: RepopackConfigMerged = mergeConfigs(fileConfig, cliConfig);
-
-  logger.trace('Merged config:', config);
-
-  // Ensure the output file is always in the current working directory
-  config.output.filePath = path.resolve(rootDir, path.basename(config.output.filePath));
-
-  const targetPath = path.resolve(directory);
-
-  const spinner = new Spinner('Packing files...');
-  spinner.start();
-
-  let packResult: PackResult;
-
-  try {
-    packResult = await pack(targetPath, config);
-  } catch (error) {
-    spinner.fail('Error during packing');
-    throw error;
+  if (options.init) {
+    await runInitCommand(cwd);
+    return;
   }
 
-  spinner.succeed('Packing completed successfully!');
-  console.log('');
-
-  if (config.output.topFilesLength > 0) {
-    printTopFiles(packResult.fileCharCounts, packResult.fileTokenCounts, config.output.topFilesLength);
-    console.log('');
-  }
-
-  printSecurityCheck(rootDir, packResult.suspiciousFilesResults);
-  console.log('');
-
-  printSummary(
-    rootDir,
-    packResult.totalFiles,
-    packResult.totalCharacters,
-    packResult.totalTokens,
-    config.output.filePath,
-    packResult.suspiciousFilesResults,
-  );
-  console.log('');
-
-  printCompletion();
+  await runDefaultCommand(directory, cwd, options);
 };
