+++ src/cli/commands/versionCommandRunner.ts
@@ -1,0 +1,6 @@
+import { getVersion } from '../../core/file/packageJsonParser.js';
+
+export const runVersionCommand = async (): Promise<void> => {
+  const version = await getVersion();
+  console.log(version);
+};
